PCExhumed: Misc code tidying

+++ source/exhumed/src/grenade.h
@@ -30,8 +30,8 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 
 void InitGrenades();
 int BuildGrenade(int nPlayer);
-void DestroyGrenade(short nGrenade);
-int ThrowGrenade(short nPlayer, int edx, int ebx, int ecx, int push1);
+void DestroyGrenade(int nGrenade);
+int ThrowGrenade(int nPlayer, int edx, int ebx, int ecx, int push1);
 void FuncGrenade(int, int, int);
 
 #endif
